# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Audience interaction API for meetups: live questions, Q\u0026A board, word clouds, feedback, presentations, QR codes and admin console. Audience requests get an anonymous participant cookie automatically.",
        "title": "Meetup API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/events/{code}/results": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "summary": "Public event results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ResultsDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "db.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "rank": {
                    "type": "integer"
                }
            }
        },
        "db.Result": {
            "type": "object",
            "properties": {
                "avg_rank": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "label": {
                    "type": "string"
                },
                "score": {
                    "description": "Score and AvgRank are only set for ranking questions.",
                    "type": "number"
                }
            }
        },
...
        "handlers.EventStatsSummaryDTO": {
            "type": "object",
            "properties": {
                "answered": {
                    "type": "integer"
                },
                "answers": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "feedback_answered": {
                    "type": "integer"
                },
                "feedback_rate": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "participants": {
                    "type": "integer"
                },
                "qa": {
                    "type": "integer"
                },
                "questions": {
                    "type": "integer"
                },
                "response_rate": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "votes": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.QADTO": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "flagged": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "voted": {
                    "type": "boolean"
                },
                "votes": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.QuestionDTO": {
            "type": "object",
            "properties": {
                "activated_at": {
                    "type": "integer"
                },
                "answered": {
                    "type": "boolean"
                },
                "correct_index": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deadline_at": {
                    "type": "integer"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_feedback": {
                    "type": "boolean"
                },
                "kind": {
                    "type": "string"
                },
                "media_type": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "my_answer": {
                    "type": "string"
                },
                "my_correct": {
                    "type": "boolean"
                },
                "my_points": {
                    "type": "integer"
                },
                "nps": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "points_base": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                },
                "prompt": {
                    "type": "string"
                },
                "respondents": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.Result"
                    }
                },
                "show_results": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "time_limit_s": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.ResultsDTO": {
            "type": "object",
            "properties": {
                "event": {
                    "$ref": "#/definitions/handlers.resultsEventDTO"
                },
                "leaderboard": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.LeaderboardEntry"
                    }
                },
                "qa": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.QADTO"
                    }
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.QuestionDTO"
                    }
                },
                "stats": {
                    "$ref": "#/definitions/handlers.EventStatsSummaryDTO"
                }
            }
        },
...
        "handlers.resultsEventDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "Admin session cookie (meetup_session) issued by POST /api/auth/login or the OIDC flow.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}